rename and move files
