+++ packages/twenty-front/src/modules/action-menu/actions/record-actions/components/RecordActionMenuEntriesSetter.tsx
@@ -1,6 +1,6 @@
+import { RecordActionMenuEntrySetterEffect } from '@/action-menu/actions/record-actions/components/RecordActionMenuEntrySetterEffect';
 import { MultipleRecordsActionMenuEntrySetterEffect } from '@/action-menu/actions/record-actions/multiple-records/components/MultipleRecordsActionMenuEntrySetterEffect';
 import { NoSelectionActionMenuEntrySetterEffect } from '@/action-menu/actions/record-actions/no-selection/components/NoSelectionActionMenuEntrySetterEffect';
-import { SingleRecordActionMenuEntrySetterEffect } from '@/action-menu/actions/record-actions/single-record/components/SingleRecordActionMenuEntrySetterEffect';
 import { WorkflowRunRecordActionMenuEntrySetterEffect } from '@/action-menu/actions/record-actions/workflow-run-record-actions/components/WorkflowRunRecordActionMenuEntrySetter';
 import { ActionViewType } from '@/action-menu/actions/types/ActionViewType';
 import { contextStoreCurrentObjectMetadataIdComponentState } from '@/context-store/states/contextStoreCurrentObjectMetadataIdComponentState';
@@ -67,14 +67,14 @@ const ActionEffects = ({
         contextStoreTargetedRecordsRule.selectedRecordIds.length === 1 && (
           <>
             {contextStoreCurrentViewType === ContextStoreViewType.ShowPage && (
-              <SingleRecordActionMenuEntrySetterEffect
+              <RecordActionMenuEntrySetterEffect
                 objectMetadataItem={objectMetadataItem}
                 viewType={ActionViewType.SHOW_PAGE}
               />
             )}
             {(contextStoreCurrentViewType === ContextStoreViewType.Table ||
               contextStoreCurrentViewType === ContextStoreViewType.Kanban) && (
-              <SingleRecordActionMenuEntrySetterEffect
+              <RecordActionMenuEntrySetterEffect
                 objectMetadataItem={objectMetadataItem}
                 viewType={ActionViewType.INDEX_PAGE_SINGLE_RECORD_SELECTION}
               />
