+++ packages/twenty-front/src/modules/action-menu/actions/record-actions/constants/DefaultSingleRecordActionsConfigV2.ts
