+++ packages/twenty-front/src/modules/action-menu/actions/record-actions/constants/WorkflowActionsConfig.ts
@@ -35,7 +35,7 @@ import {
   IconTrashX,
 } from 'twenty-ui';
 
-export const WORKFLOW_SINGLE_RECORD_ACTIONS_CONFIG: Record<
+export const WORKFLOW_ACTIONS_CONFIG: Record<
   string,
   ActionMenuEntry & {
     actionHook: ActionHook;
